fix: update migration attributes

+++ database/migrations/2023_11_11_080623_create_cars_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('cars', function (Blueprint $table) {
             $table->id('vehicle_id')->unique(); // Unique constraint for shared primary key
             $table->string('fuel_type');
-            $table->integer('trunk_size');
+            $table->float('trunk_size');
             $table->timestamps();
         });
     }
